refactor: merge calendar files
