+++ utils/index.ts
@@ -1,7 +1,6 @@
-import * as calendar from './calendar'
 import { logger } from './logs'
 import * as sql from './sql'
 import * as format from './format'
 import * as object from './object'
 
-export { calendar, logger, sql, format, object }
+export { logger, sql, format, object }
